+++ src/main/java/com/atwoz/survey/application/membersurvey/SoulmatePolicy.java
@@ -1,4 +1,4 @@
-package com.atwoz.survey.application.membersurvey.soulmatepolicy;
+package com.atwoz.survey.application.membersurvey;
 
 import com.atwoz.survey.infrastructure.membersurvey.dto.SurveySoulmateResponse;
